Make dora rerun easier to read
