+++ node-hub/dora-rerun/README.md
@@ -33,6 +33,28 @@ cargo install --git https://github.com/dora-rs/dora dora-rerun
 - image: UInt8Array + metadata { "width": int, "height": int, "encoding": str }
 - boxes2D: StructArray + metadata { "format": str }
 - text: StringArray
+- jointstate: Float32Array
+
+## (Experimental) For plotting 3D URDF
+
+```bash
+pip install git+https://github.com/rerun-io/rerun-loader-python-example-urdf.git
+```
+
+Make sure to name the dataflow as follows:
+
+```yaml
+- id: rerun
+  path: dora-rerun
+  inputs:
+    jointstate_<ENTITY_NAME>: <ENTITY_NAME>/jointstate
+  env:
+    <ENTITY_NAME>_urdf: /path/to/<ENTITY_NAME>.urdf
+    <ENTITY_NAME>_transform: 0 0.3 0
+```
+
+> [!IMPORTANT]  
+> Make sure that the urdf file name correspond to your dataflow object name otherwise, it will not be able to link to the corresponding entity.
 
 ## Configurations
