+++ node-hub/dora-rerun/pyproject.toml
@@ -4,3 +4,7 @@ build-backend = "maturin"
 
 [project]
 name = "dora-rerun"
+dependencies = [
+    'rerun_sdk==0.19.1',
+    "git+https://github.com/rerun-io/rerun-loader-python-example-urdf.git",
+]
