+++ node-hub/dora-rerun/src/main.rs
@@ -1,6 +1,7 @@
 //! Demonstrates the most barebone usage of the Rerun SDK.
 
 use std::{collections::HashMap, env::VarError};
+pub mod urdf;
 
 use dora_node_api::{
     arrow::{
@@ -10,101 +11,11 @@ use dora_node_api::{
     DoraNode, Event, Parameter,
 };
 use eyre::{eyre, Context, ContextCompat, Result};
-use k::Chain;
 use rerun::{
     components::{ImageBuffer, RotationAxisAngle},
     external::re_types::ArrowBuffer,
-    Angle, ImageFormat, RecordingStream, Rotation3D, SpawnOptions, Text, Vec3D,
+    Angle, ImageFormat, Rotation3D, SpawnOptions, Text,
 };
-pub struct MyIntersperse<T, I> {
-    iterator: I,
-    sep: T,
-    nxt: Option<T>,
-}
-
-pub trait MyIntersperseExt<T: Clone, I: Iterator>: Iterator<Item = T> {
-    fn my_intersperse(self, sep: T) -> MyIntersperse<T, I>;
-}
-
-impl<T: Clone, I: Iterator<Item = T>> MyIntersperseExt<T, I> for I {
-    fn my_intersperse(mut self, sep: T) -> MyIntersperse<T, I> {
-        let next = self.next();
-        MyIntersperse {
-            iterator: self,
-            sep: sep.clone(),
-            nxt: next,
-        }
-    }
-}
-
-impl<T: Clone, I: Iterator<Item = T>> Iterator for MyIntersperse<T, I> {
-    type Item = T;
-    fn next(&mut self) -> Option<Self::Item> {
-        if let Some(item) = self.nxt.take() {
-            Some(item)
-        } else {
-            self.nxt = self.iterator.next();
-            if self.nxt.is_some() {
-                Some(self.sep.clone())
-            } else {
-                None
-            }
-        }
-    }
-}
-
-fn get_entity_path(link: &k::Node<f32>, urdf_path: &str) -> String {
-    let mut ancestors: Vec<_> = link
-        .iter_ancestors()
-        .map(|node| node.link().as_ref().unwrap().name.clone())
-        .collect();
-    ancestors.push(String::from(urdf_path));
-    ancestors
-        .into_iter()
-        .rev()
-        .my_intersperse(String::from("/"))
-        .collect()
-}
-
-fn update_visualization(
-    chain: &Chain<f32>,
-    rec: &RecordingStream,
-    id: &str,
-    positions: &[f32],
-) -> Result<()> {
-    chain.set_joint_positions_clamped(&positions);
-
-    chain.update_transforms();
-    chain.update_link_transforms();
-
-    for link_name in chain.iter_links().map(|link| link.name.clone()) {
-        let link = chain.find_link(&link_name).unwrap();
-        let entity_path = get_entity_path(&link, &id);
-        let link_to_world = link
-            .world_transform()
-            .context("Could not get world transform")?;
-        let link_to_parent = if link_name != "base_link" {
-            let parent = link.parent().unwrap();
-            parent.world_transform().unwrap().inv_mul(&link_to_world)
-        } else {
-            link_to_world
-        };
-        let link_to_parent_mat = link_to_parent.to_matrix();
-
-        let trans = link_to_parent_mat.column(3);
-        let trans = trans.as_slice();
-        let quat = link_to_parent.rotation.quaternion();
-        rec.log(
-            entity_path,
-            &rerun::Transform3D::from_translation_rotation(
-                Vec3D::new(trans[0], trans[1], trans[2]),
-                rerun::Quaternion::from([quat.i, quat.j, quat.k, quat.w]),
-            ),
-        )
-        .context("Could not log transform")?;
-    }
-    Ok(())
-}
 
 fn main() -> Result<()> {
     // rerun `serve()` requires to have a running Tokio runtime in the current context.
@@ -345,7 +256,7 @@ fn main() -> Result<()> {
                     }
                 }
 
-                update_visualization(&chain, &rec, &id, &positions)?;
+                urdf::update_visualization(&chain, &rec, &id, &positions)?;
             }
         }
     }
